+++ src/mantle/testing/doubles/class-spy-rest-server.php
@@ -2,17 +2,32 @@
 
 namespace Mantle\Testing\Doubles;
 
+use WP_REST_Request;
 use WP_REST_Server;
 
+/**
+ * Spy REST Server
+ *
+ * A spy class for WP_REST_Server that allows us to inspect the headers and body
+ * of the response without sending it to the client.
+ */
 class Spy_REST_Server extends WP_REST_Server {
 
-	public array $sent_headers  = [];
-	public $sent_body;
-	public $last_request;
-	public $override_by_default = false;
+	/**
+	 * @var array<string, string>
+	 */
+	public ?array $sent_headers  = [];
+
+	public ?int $sent_status = null;
+
+	public ?string $sent_body = null;
+
+	public ?WP_REST_Request $last_request = null;
+
+	public bool $override_by_default = false;
 
 	/**
-	 * Gets the raw $endpoints data from the server.
+	 * Gets the raw endpoints data from the server.
 	 *
 	 * @return array
 	 */
@@ -28,7 +43,16 @@ class Spy_REST_Server extends WP_REST_Server {
 	 * @return mixed
 	 */
 	public function __call( $method, $args ) {
-		return call_user_func_array( [$this, $method], $args );
+		return call_user_func_array( [ $this, $method ], $args );
+	}
+
+	/**
+	 * Sends an HTTP status code.
+	 *
+	 * @param int $code HTTP status.
+	 */
+	protected function set_status( $code ) {
+		$this->sent_status = $code;
 	}
 
 	/**
@@ -58,6 +82,7 @@ class Spy_REST_Server extends WP_REST_Server {
 	 */
 	public function dispatch( $request ) {
 		$this->last_request = $request;
+
 		return parent::dispatch( $request );
 	}
 
@@ -87,5 +112,15 @@ class Spy_REST_Server extends WP_REST_Server {
 		$result          = parent::serve_request( $path );
 		$this->sent_body = ob_get_clean();
 		return $result;
+	}
+
+	/**
+	 * Clear the stored response data for the spy.
+	 */
+	public function reset_spy(): void {
+		$this->sent_headers = null;
+		$this->sent_status  = null;
+		$this->sent_body    = null;
+		$this->last_request = null;
 	}
 }